smartlist: fix selection
RecentModel emits a dataChanged every second for updating call timers.
This was causing scrolling issues in the list. We don't connect to dataChanged
from our NSTreeController but from the View directly. The reason is that there
is no way to update a row in a NSTreeController, only insert/remove/move and
this is causing selection updates.

Issue: #81198
Change-Id: Ib347ff5a69f901da2bdf1d0b3a3d368fdfd23c1c