history: fix history limit
Since history is now handled by LRC, set the history limit instead
of storing it client side.

There is a race condition in LRC, we need to make sure daemon is
started before creating the history, otherwise dring.yml is not
parsed when we ask what is the currently set historyLimit and returns
default value.

Change-Id: I433a37ab559721df82225a6fe2695525ed85fcd0
Tuleap: #103