persons: improve caching
The caching implemented in the ImageManipulationDelegate is preventing
the change of anycontact photo at runtime (even the profile one).

This commit monitors the changes on a photo and invalidates the cache
when a changes is detected.
The caching system was also not safe to use because it was building
indexes out of possibily invalid persons.
The BrokerVC and PersonLinkerVC now use the DecorationRole to increase
photo quality.

Change-Id: I1a364bbb0ade130868014a10ec127eff8a7e620b
Tuleap: #697