Update patch set 4
Patch Set 4:

Just found out, when choosing a password for a new Ring account, I can enter a three character password, switch to the confirmation field, type the same three chars and the checkmark for the confirmation appears but not the one for the password. And the OK button stays grey. I think a small popover explaining password validity could help.

Patch-set: 4