Update patch set 6
Patch Set 6: Code-Review-1

It seems to work fine for the content of this patch.
However, I'd like to talk to you about the cancel action during the process of linking. Each time I cancel a linking, the created account remains stored. So, when I try to perform the operation again, I see in the logs "Updating all accounts" and the list of all the previous created accounts, which are duplicated.
Do you prefer to check this here or to move it to an issue plus a new patchset ? If you can do something and if this is reproductible of course.

Patch-set: 6
Label: Code-Review=-1