Update patch set 6
Patch Set 6:

Again, not much I can do in the context of this patch. I have removed the Cancel action, and I delete accounts in case of error (if LRC does not do it). There is a small window of time where you can access the main UI but I can't do anything about it.

Patch-set: 6
Label: Code-Review=0