ui: update layout constraints
There was an issue with the wizard's help popover content. Some layout
contraints were not interpreted by the 10.10 SDK (used for packaging)
Text is now wrapping properly, displaying all the content.

Another issue was with the welcome page message. The uri field was
compressed because the french translation takes a lot more space.
The welcome message is now allowed to take more spaces horizontally,
allowing the uri to be properly displayed.

Change-Id: I65a6ecf1b4f933fd1b3804680657f1afd17fe692
Tuleap: #1353