smartlist: change first responder
Switch between Smart, contacts and history list can be hard. User have
to set focus on the smartlist component in order to activate the menu
items in the top screen bar under the 'Show' button.

By changing the first responder to the Ring window, this items are
always enabled. Actions are then forwarded to the panel view controller

This will be usefull for trust request because we will need to display
the contact list associated with an account

Change-Id: I02baeb408828f36b098a1d8d22b53d252213b6e2
Tuleap: #1553