SmartList: fix ContactMethod duplication
When searching for new number two ContactMethods are created. This patch
fixes it by using other method from PhoneDirectoryModel to create new
ContactMethod(instead of using only URL now it uses URL and account). Also
this patch make connection between ContactMethod and account in case if it
was not associated before.

Change-Id: I231224506575dad2d11e48e36aef27bfb53ab2db
Reviewed-by: Anthony Léonard <anthony.leonard@savoirfairelinux.com>