SmartList: one entry for Ring and Mac contact
If MAC Address Book is available when user save contact two persons
are created. It may be unclear for user because those entries looks
similar and point to the same ringID. And it could cause other duplication
in future if user will select contact related to Address Book to make
call. This patch use same person for Ring and Mac contact, it specifies
collection to save person when user create new contact and also it checks
if person exists before adding new from Mac contacts.

Change-Id: I80daccba1bcbcfc8eed27b6962e692171012fdd9