smartlist: do not display bestID if equal to name
In the case where the name of a contact is exactly the same as its
registered name or RingID, we were still displaying both. It happens
when a contact has just been added.

Now, we hide the bestID if it is equal to the name and show it back
on name changes for its security purpose.

Change-Id: I743df652661358a7a2940844c6c3770ebcaf29e6