im: remove bestID if equal to bestName
In order to reduce information redundancy and possible confusion, the
bestID of the contact we are currently talking to is hidden if it is
equal to its bestName. This one shows a different name only if we
received the contact VCard and its profile name is different than its
registered name.

Moreover, in case we only display the bestName, we center it
vertically for a cleaner appearance.

Change-Id: I29aa32d4c8394fded057c18f2b8c5446d3cfa11f