modernize avatar placeholder
Default avatar displayed for new contacts is replaced by a more modern
color circle with a capital letter inside.

The color palette used is the same as in the GNOME client. It comes
from the material.io website and is a 16 color subset of the full one
proposed by Google.

The letter is the first one of the bestName() corresponding to the
current item.

Change-Id: Ib5b83c76e73ee8af658b0515b317e4f8ce82be27