account: add password change modal in settings
A button is added in the Ring account setting panel which trigger a
sliding modal enabling a user to change the password used to encrypt
the master key archive (its Ring “account”).

To prevent mistakes, the old password is not required if we know that
there isn't any. The new password is also asked twice for confirmation.

Change-Id: Ibbc0a87f1d108b7fdb66382e8f8f491461eeaccf