build: ship QtSql and SQLite driver with bundle
Both of them are required in order to use the new LRC models.

The QtSql plugin is added as a new framework inside the .app bundle.
It requires the QSQLite driver in order to use this database as a
backend. Thus it is also included.

libsqlite3.dylib itself is not required as it is already available in
the base system.

Change-Id: I4f5236369b121d1eddf7bbc469496dda660a5862
Reviewed-by: Olivier Soldano <olivier.soldano@savoirfairelinux.com>