draw picture for new lrc model of conversation
The conversationPhoto() from PixmapManipulator is implemented in order
to provide the picture associated to a conversation. In case no
picture is present in database, we draw the fallback avatar.

With this implementation, picture for conference will be the one of
the first participant in the conversation. This limitation is left
for future improvement.

Change-Id: I6be58465d48bfe7a32e4c19b4aca50a077030324
Reviewed-by: Guillaume Roguez <guillaume.roguez@savoirfairelinux.com>