reimplement CurrentCallVC with new call model
This patch reimplements CurrentCallVC using the call model introduced
recently in LRC.

As of this patch, some minor features are missing such as conversation
time or contact picture display when receiving or starting a call.
Call panel may also appear and disappear many times during connection
establishment. All of that is meant to be fixed very soon and doesn't
prevent starting or receiving a call.

Change-Id: Ie51f75ffc267e955679a4cc75663f92e03c36eef
Reviewed-by: Kateryna Kostiuk <kateryna.kostiuk@savoirfairelinux.com>