fix account selection after first account creation
Account wasn't properly set due to a not handled exception. This is no
longer the case as the first account of the list is now selected by
default when no account is yet saved in user configuration.

Moreover, having an empty AccountModel shouldn't cause any crash as
this case is now handled with exceptions.

Change-Id: Ibb3e386d80e946b2507916b3e44cb28af62bc3c7