improve name display in call view
Both incoming call view and in-call view were displaying the ringID of
the contact instead of its best name. This is now fixed and more user
friendly.

Plus, two non-necessary constraints in incoming call view are removed.
This gives a better aligment of textfields when using a language other
than english.

Change-Id: I675c95b5f996b3fa8853554f14c0ed0aab44be87