hide list type selector when no requests in list
The whole selector is hidden when no requests are pending for a given
account. In this case, only conversation list is displayed.

When it needs to be visible again, we re-add it to the view with its
constraints.

Also diminish potential number of data reloads in smartview by moving
some code.

Change-Id: Ib00aa8b59a70e65015f9c28402278054b591522d
Reviewed-by: Kateryna Kostiuk <kateryna.kostiuk@savoirfairelinux.com>