fix account selection on fresh install
Fix bug where, at first user creation, Ring was crashing at the end of
the wizard just before displaying the main user interface. This was
caused by a preference string set to nil on first start.

Now, when to string is nil, we select the first account of the list as
this should happen only at first account creation (thus returning the
right one).

Change-Id: Id3a2c4153d920611a3cc0cbd694f8d96d772330a