fix unread msg counter update + dock icon counter
2 problems were preventing the unread msg counter to work properly:
- Messages were not marked as read when displayed.
- Signals announcing status update were not listened to.

Moreover, the one present on dock icon was still listening to old LRC
mechanisms and is thus removed (until further improvements).

Finally, as the interactionStatusUpdated signal is triggered multiple
times when many messages are unread, some logic is added to debounce
it so only one refresh is triggered for a group of update.

Change-Id: Ieaa02d685bbd85d965de5315e8694ccdbd18eff1