fix window size changes during some message interactions
After a call, a signal wasn't disconnected properly. This was leading
to resizings for no apparent reason when receiving or sending a
message after this was ended.

This was caused by the in-call chat window, invisible at this time, to
be opened because of that signal. We now disconnect it when closing
call view.

Change-Id: Icd211bad795e146353e099004c1ba13eadccf9ba