smart list: fix crash when removing conversation
When the client receives signal conversationRemoved it tries to
reload filtered conversations. At this point, filtered conversations
are not updated yet and it could cause a crash. This reloading
is unnecessary because after the conversation is removed client receives
a modelChanged signal where smart list data should be reloaded.

Gitlab: #272

Change-Id: Ia9baaa087e4ce56b40d79bbc8a9356b403f02ae4